+++ src/Telegram.Bot/Exceptions/ApiExceptionParser.cs
@@ -11,9 +11,16 @@ namespace Telegram.Bot.Exceptions
             new BadRequestExceptionInfo<ChatNotFoundException>("chat not found"),
             new BadRequestExceptionInfo<UserNotFoundException>("user not found"),
             new BadRequestExceptionInfo<InvalidUserIdException>("USER_ID_INVALID"),
+
+            #region Stickers
+
             new BadRequestExceptionInfo<InvalidStickerSetNameException>("sticker set name invalid"),
             new BadRequestExceptionInfo<InvalidStickerEmojisException>("invalid sticker emojis"),
             new BadRequestExceptionInfo<InvalidStickerDimensionsException>("STICKER_PNG_DIMENSIONS"),
+            new BadRequestExceptionInfo<StickerSetNameExistsException>("sticker set name is already occupied"),
+
+            #endregion
+
             new BadRequestExceptionInfo<ContactRequestException>("phone number can be requested in a private chats only"),
 
             new ForbiddenExceptionInfo<ChatNotInitiatedException>("bot can't initiate conversation with a user"),